+++ js/haskier.js
@@ -1,31 +1,137 @@
 'use strict';
 
-var hidden = $('#terminal, #infos, #fullscreen_trigger').hide();
+var hidden = $('#terminal, #fullscreen_trigger').hide(),
+  dontRecoverPrompt = false, hsf_conds = [], todo = [], justRestored = false, justRestoredLabel = false,
+  ignoreKeys = false;
+
+/*const*/var readingLineDuration = 50;
+const backup_prefix = 'hbak_auto_';
 
 /**
   * Start the game !
   */
-function go() {
+function ready() {
   hidden.show();
+
+  game.event('label', function() {
+    if(game.getVar('FREEZE_UNTIL_TODO'))
+      game.setVar('dont_update_prompt', true);
+
+    if(!justRestoredLabel)
+      saveGame(); // useless ?
+    else
+      justRestoredLabel = false;
+  });
+
+  game.event('display', function(text, i, code) {
+    if(game.getVar('human_talking'))
+      scope.human(game.getVar('human_talking') + ' :  ' + text);
+    else {
+      ignoreKeys = true;
+      term.set_prompt('');
+      display(game.getVar('ALLOW_TRANSLATION') ? tr(text) : text);
+      setTimeout(function() {
+        ignoreKeys = false;
+        if(!game.getVar('dont_update_prompt'))
+          updatePrompt();
+        go();
+      }, readingLineDuration * text.length);
+    }
+  });
+
+  game.event('include', function(filename) {
+    if(!HSF_files[filename])
+      console.error('Scenaristic file /hsf/' + filename + ' was not found');
+
+    return HSF_files[filename];
+  });
+
+  go();
 }
+
+/**
+  * Run the next game's instruction
+  */
+var go = function() {
+  /*if(game.finished())
+    return ;*/
+
+  game.step();
+};
 
 /**
   * Save the game
   */
 function saveGame() {
+  if(!saveSupport)
+    return console.warn('saveGame() was ignored because localStorage doesn\'t work');
+
+  delete vars.scope;
+
   try {
+    var lastSave = localStorage.getItem('haskier');
+
     localStorage.setItem('haskier', JSON.stringify({
       view    : term.export_view(),
       vars    : vars,
       time    : (new Date()).getTime(),
       data    : save.data,
-      server  : server.export()
+      server  : server.export(),
+      scope   : game.diffScope(),
+      label   : game.label(),
+      marker  : game.marker(),
+      dsas    : didSomethingAfterSave
     }));
+
+    // Backup the previous save
+    if(lastSave && localStorageFree() > lastSave.length + 16) {
+      var id = 1, name;
+
+      while(localStorage.getItem(name = backup_prefix + id))
+        id++;
+
+      try { localStorage.setItem(name, lastSave); }
+      catch(e) { console.error('Failed to backup last Haskier save\n' + e.stack); }
+    } else if(lastSave) {
+      // If there is not enough memory to backup the last save
+      var keys = Object.keys(localStorage), backups = [], taken = 0;
+
+      // This loop permit to sort the backups ID
+      for(var i = 0; i < keys.length; i++) {
+        if(keys[i].substr(0, backup_prefix.length) === backup_prefix) {
+          backups.push(keys[i].substr(backup_prefix.length));
+          taken += localStorage.getItem(keys[i]).length;
+        }
+      }
+
+      if(taken > lastSave.length + 16) {
+        backups = backups.sort();
+
+        for(i = 0; i < backups.length; i++) {
+          localStorage.removeItem(backup_prefix + i);
+
+          if(localStorageFree() > lastSave.length + 16)
+            break;
+        }
+
+        console.info(i + ' backups were deleted to make free space for backuping the last save (' + (lastSave.length / 1024).toFixed(1) + ' Kb)');
+        try { localStorage.setItem(backup_prefix + (parseInt(backups[backups.length - 1]) + 1), lastSave); }
+        catch(e) { console.error('Failed to backup last Haskier save\n' + e.stack); }
+      } else {
+        if(backups.length)
+          console.warn('There is no enough total space to backup last save');
+        else
+          console.warn('Even if deleting ' + backups.length + ' backups, there is not enough space to backup last save');
+      }
+    }
   }
 
   catch(e) {
     alert(tr('Failed to save game. Please try again. If that doesn\'t work, type `force-save` in the terminal and press Return.'));
+    console.error('Failed to save game\n' + e.stack);
   }
+
+  vars.scope = scope;
 }
 
 /**
@@ -75,18 +181,33 @@ function treatQueue() {
   if(runningCmd)
     return report_bug('treatQueue() has been called but a command is already running');
 
-  term.pause(); // $('#cover').show(); -> doesn't work better
+  ignoreKeys = true; //term.pause(); // $('#cover').show(); -> doesn't work better
 
   var cmd    = queue.splice(0, 1)[0];
   runningCmd = true;
 
   exec(cmd, function() {
-    term.resume(); // $('#cover').hide(); -> doesn't work better
+    ignoreKeys = false; //term.resume(); // $('#cover').hide(); -> doesn't work better
     runningCmd = false;
+    didSomethingAfterSave = true;
     saveGame();
 
     if(queue.length)
       treatQueue();
+    else if(todo.length) {
+      // Here, we've wait queue is empty to check if todo is accomplished
+      // And because condition contains 'todo.length' we know that todo list is not empty
+
+      for(var i = 0; i < todo.length; i++) {
+        if(window.eval(todo[i])) {
+          todo.splice(i, 1);
+          i--;
+        }
+      }
+
+      if(!todo.length)
+        go();
+    }
   });
 }
 
@@ -209,14 +330,15 @@ function parseCommand(cmd) {
 function updateUI() {
   updatePrompt();
   $('.terminal, .cmd').css('font-family', (save.data.font ? save.data.font + ', ' : '') + 'Consolas, Courier, "Inconsolata"');
-  $('#infos')[save.data.hideInfobar ? 'hide' : 'show']();
+  //$('#infos')[save.data.showInfobar ? 'show' : 'hide']();
 }
 
 /**
   * Update the terminal's prompt
   */
 function updatePrompt() {
-  term.set_prompt(format('${green:${name}}:${blue:' + server.chdir() + '}$ '));
+  if(game && !game.getVar('dont_update_prompt'))
+    term.set_prompt(format('${green:${name}}:${blue:' + server.chdir() + '}$ '));
 }
 
 var queue      = [];           // Commands' queue
@@ -233,13 +355,21 @@ var term = $('#terminal').terminal(function(cmd, term) {
   // If there is a catch callback
   if(catchCommand) {
     // We store it in memory...
-    var callback = catchCommand;
+    var callback = catchCommand, ret;
     // To delete it in the variable (this permit to remove some bugs)
     catchCommand = null;
     // If the callback does no return 'true'
-    if(callback(cmd) !== true)
+    if((ret = callback(cmd)) !== true && !dontRecoverPrompt)
       // We recover the prompt
       updatePrompt();
+    else {
+      //if(ret === true)
+        dontRecoverPrompt = false;
+
+      if(!catchCommand)
+        // We recover the catcher
+        catchCommand = callback;
+    }
   } else
     // Else, we run the command
     command(cmd);
@@ -248,6 +378,25 @@ var term = $('#terminal').terminal(function(cmd, term) {
   name         : 'haskier-terminal',
   prompt       : '$ ',
   tabcompletion: true,
+  keydown      : function(e) {
+    if(ignoreKeys)
+      return false;
+
+    // If there is a catch callback
+    if(keydownCallback) {
+      // We store it in memory...
+      var callback = keydownCallback, ret;
+      // To delete it in the variable (this permit to remove some bugs)
+      keydownCallback = null;
+      // If the callback does returns 'true'
+      if(callback(e) === true)
+        // Restore the callback
+        keydownCallback = callback;
+
+      // Prevent the key event
+      return false;
+    }
+  },
   completion   : function(term, word, callback) {
     // If there is a catch callback
     if(catchCommand) {
@@ -300,15 +449,14 @@ $('#fullscreen_trigger').on('click', function() {
 });
 
 // Load game save
-var saveSupport = false; // Is save system supported ?
-var save, is_save;       // Is there a valid save ?
+var saveSupport = hasLocalStorage && localStorageWorking;
+  // Is save system supported ?
+var save, is_save; // Is there a valid save ?
 
-if(typeof localStorage !== 'object' || typeof localStorage.setItem !== 'function' || typeof localStorage.getItem !== 'function' || typeof localStorage.removeItem !== 'function' || typeof localStorage.clear !== 'function')
+if(!saveSupport)
   alert(tr('Your browser doesn\'t support localStorage feature. Your will not be able to save your game.\nTo save your progression, please use a newer browser or update this one.'));
 else {
-  saveSupport = true;
-
-  try {localStorage.setItem('__localStorage_test', 'abcdefghij'.repeat(5000)); }
+  try {localStorage.setItem('__localStorage_test', 'abcdefghij'.repeat(25000) /* 250 Kb */); }
   catch(e) { alert(tr('localStorage test has failed. You will perhaps not be able to save your game.')); }
 
   localStorage.removeItem('__localStorage_test');
@@ -323,7 +471,7 @@ else {
         localStorage.removeItem('haskier');
       } else {
         is_save = true;
-        console.info('A (valid) Haskier save was found');
+        console.info('A valid Haskier save was found | ' + (new Date(save.time)).toString());
       }
     }
   }
@@ -343,6 +491,9 @@ if(!is_save) {
   vars['$c_red'   ] = '#FE1B00';
   vars['$c_purple'] = '#8066B3';
 
+  vars.haskier  = '${f_cyan,italic:Haskier}';
+  vars.hypernet = '${f_#8066B3,bold,italic:HyperNet}';
+
   // Initialize the terminal
   term.clear();
 } else {
@@ -353,7 +504,7 @@ if(!is_save) {
   // Fix a bug with jQuery.terminal plugin
   save.view.interpreters = term.export_view().interpreters;
   // Import view...
-  term.import_view(save.view);
+  //term.import_view(save.view);
 }
 
 // Update interface
@@ -362,30 +513,199 @@ updateUI();
 // Define global HSF scope
 var scope = {
   display: function(text) {
-    display(text);
+    display(game.getVar('ALLOW_TRANSLATION') ? tr(text) : text);
+    go();
+  },
+
+  human: function(text) {
+    display(game.getVar('ALLOW_TRANSLATION') ? tr(text) : text, go);
   },
 
   clear: function() {
-    term.clear();
+    term.clear(); go();
+  },
+
+  question: function(msg) {
+    dontRecoverPrompt = true;
+    question(game.getVar('ALLOW_TRANSLATION') ? tr(msg) : msg, function(answer) {
+      // Question callback
+      game.setVar('answer', answer);
+      dontRecoverPrompt = null;
+      go();
+    });
+  },
+
+  choice: function() {
+    dontRecoverPrompt = true;
+    term.echo(' ');
+
+    if(game.getVar('ALLOW_TRANSLATION')) {
+      for(var i = 0; i < arguments.length; i++)
+        arguments[i] = tr(arguments[i]);
+    }
+
+    choice(arguments, function(answer) {
+      // Choice callback
+      dontRecoverPrompt = null;
+      term.echo(' ');
+      game.setVar('answer', answer);
+      go();
+    });
+  },
+
+  confirm: function(msg) {
+    dontRecoverPrompt = true;
+    term.echo(' ');
+    confirm(game.getVar('ALLOW_TRANSLATION') ? tr(msg) : msg, function(answer) {
+      // Confirm callback
+      dontRecoverPrompt = null;
+      term.echo(' ');
+      game.setVar('answer', answer);
+      go();
+    });
+  },
+
+  leave: function(dontGo) {
+    // Permit prompt to refresh again, update it and permit user to access the terminal
+    game.setVar('dont_update_prompt', false);
+    updatePrompt();
+    catchCommand = false;
+
+    if(!dontGo)
+      go();
+  },
+
+  freeze: function() {
+    // If 'FREEZE_UNTIL_TODO' var is set to true, we make prompt unable to refresh
+    game.setVar('dont_update_prompt', true);
+    go();
+  },
+
+  restore: function(dontRestoreScope) {
+    // Restore the checkpoint
+    if(save.label) {
+      justRestoredLabel = true;
+
+      if(!game.goLabel(save.label)) {
+        console.error('Failed to go to label "' + save.label + '"');
+        justRestoredLabel = false;
+      } else {
+        if(save.dsas) {
+          game.pass();
+
+          while(game.current(1)) {
+            if(game.current(1).js && game.current(1).js.match(/^todo *\(/))
+              break;
+
+            game.pass();
+          }
+        }
+
+        term.import_view(save.view);
+      }
+
+      justRestored = true; // put it here or in {if(game.goLabel) ???}
+    }
+
+    // If the save contains the script's diff scope, restore it
+    if(save.scope && !dontRestoreScope) {
+      var keys = Object.keys(save.scope);
+
+      for(var i = 0; i < keys.length; i++)
+        scope[keys[i]] = save.scope[keys[i]];
+    }
+
+    go();
+  },
+
+  server: server,
+  save  : save,
+  game  : game,
+
+  todoModels: {},
+  setTodoModel: function(name, model) {
+    scope.todoModels[name] = model;
+    go();
+  },
+  todo: function() {
+    // todo = []; // useless here
+
+    // If 'FREEZE_UNTIL_TODO' var is set to true, we make user able again to use terminal
+    // ... because it was frozen
+    if(game.getVar('FREEZE_UNTIL_TODO'))
+      scope.leave(true);
+
+    // Prepare the todo list...
+    for(var i = 0; i < arguments.length; i++) {
+      if(typeof arguments[i] === 'string')
+        todo.push(scope.todoModels[arguments[i]]);
+      else
+        todo.push(arguments[i]);
+    }
+
+    if(!justRestored) {
+      didSomethingAfterSave = false;
+      saveGame();
+    } else justRestored = false;
+  },
+
+  goto: function(label) {
+    game.goLabel(label);
+    go();
+  },
+
+  repeat: function(label) {
+    game.repeatLabel();
+    go();
+  },
+
+  next: function(label) {
+    game.nextLabel();
+    go();
+  },
+
+  wait: function(time) {
+    if(typeof time === 'string') {
+      display(time);
+      keydownCallback = go;
+    } else {
+      ignoreKeys = true;
+      setTimeout(function() {
+        ignoreKeys = false;
+        go();
+      }, time);
+    }
+  },
+
+  localStorageStats: function() {
+    scope.answer = localStorageStats();
+    go();
   }
 };
 
+vars.scope = scope;
+
 // Define some callbacks
-var afterCommand, catchCommand;
+var afterCommand, catchCommand, keydownCallback;
 // Load game
-var serverGame, game;
+var serverGame, game, didSomethingAfterSave = false, HSF_files = {};
 
 $.ajax({
-  url: 'com/get-game.run',
+  url     : 'com/get-game.run',
   dataType: 'json',
-  success: function(data) {
+  cache   : false,
+  success : function(data) {
     serverGame = data;
-    game = HSF.parse(data['scenario.hsf'], true);
 
-    if(typeof game === 'string')
-      alert(tr('Received data are not valid. Please try later.\n\nDetails :\n${err}', [game]));
-    else
-      go();
+    // Parse all scenaristic files
+    var filenames = Object.keys(data.hsf);
+
+    for(var i = 0; i < filenames.length; i++)
+      HSF_files[filenames[i]] = HSF.parse(data.hsf[filenames[i]], scope);
+
+    game = HSF_files['main.hsf'];
+
+    ready();
   },
   error: function(err) {
     alert(tr('Failed to load game. Please refresh the page.'));